Fix Seedless onboarding
