+++ packages/core-mobile/app/seedless/services/SeedlessService.ts
@@ -3,6 +3,8 @@ import Logger from 'utils/Logger'
 import { ACCOUNT_NAME } from '../consts'
 import { SeedlessSessionManager } from './storage/SeedlessSessionManager'
 import SeedlessSession from './SeedlessSession'
+import { SeedlessPubKeysStorage } from './storage/SeedlessPubKeysStorage'
+import { transformKeyInfosToPubKeys } from './wallet/transformKeyInfosToPubkeys'
 
 /**
  * Service for cubesigner-sdk
@@ -26,6 +28,22 @@ class SeedlessService {
 
   init({ onSessionExpired }: { onSessionExpired: () => void }): void {
     this.session.setOnSessionExpired(onSessionExpired)
+  }
+
+  async refreshPublicKeys(): Promise<void> {
+    try {
+      const storedPubKeys = await SeedlessPubKeysStorage.retrieve()
+      if (storedPubKeys.length === 0) {
+        const allKeys = await this.getSessionKeysList()
+
+        const pubKeys = transformKeyInfosToPubKeys(allKeys)
+        Logger.info('saving public keys')
+        await SeedlessPubKeysStorage.save(pubKeys)
+      }
+    } catch (error) {
+      Logger.error(`Unable to save public keys`, error)
+      throw new Error(`Unable to save public keys`)
+    }
   }
 
   /**
